Refactor: split ScadaSettings into ProactorSettings and ScadaSettings(ProactorSettings)
