+++ gw_spaceheat/proactor/__init__.py
@@ -15,6 +15,7 @@ Particular questions:
 """
 
 from proactor.logger import ProactorLogger
+from proactor.config import ProactorSettings
 from proactor.message import Header
 from proactor.message import Message
 from proactor.proactor_implementation import MQTTCodec
@@ -28,6 +29,7 @@ from proactor.sync_thread import AsyncQueueWriter
 from proactor.sync_thread import SyncAsyncInteractionThread
 from proactor.sync_thread import SyncAsyncQueueWriter
 
+
 __all__ = [
     "AsyncQueueWriter",
     "Communicator",
@@ -38,6 +40,7 @@ __all__ = [
     "MQTTCodec",
     "Proactor",
     "ProactorLogger",
+    "ProactorSettings",
     "Runnable",
     "ServicesInterface",
     "SyncAsyncInteractionThread",
