+++ gw_spaceheat/proactor/config/__init__.py
@@ -1,20 +1,21 @@
 """Settings for the GridWorks Scada, readable from environment and/or from env files."""
-from .logging import DEFAULT_BYTES_PER_LOG_FILE
-from .logging import DEFAULT_FRACTIONAL_SECOND_FORMAT
-from .logging import DEFAULT_LOG_FILE_NAME
-from .logging import DEFAULT_LOGGING_FORMAT
-from .logging import DEFAULT_NUM_LOG_FILES
-from .logging import FormatterSettings
-from .logging import LoggerLevels
-from .logging import LoggingSettings
-from .logging import RotatingFileHandlerSettings
-from .mqtt import MQTTClient
-from .paths import DEFAULT_BASE_DIR
-from .paths import DEFAULT_BASE_NAME
-from .paths import DEFAULT_LAYOUT_FILE
-from .paths import DEFAULT_NAME
-from .paths import DEFAULT_NAME_DIR
-from .paths import Paths
+from proactor.config.logging import DEFAULT_BYTES_PER_LOG_FILE
+from proactor.config.logging import DEFAULT_FRACTIONAL_SECOND_FORMAT
+from proactor.config.logging import DEFAULT_LOG_FILE_NAME
+from proactor.config.logging import DEFAULT_LOGGING_FORMAT
+from proactor.config.logging import DEFAULT_NUM_LOG_FILES
+from proactor.config.logging import FormatterSettings
+from proactor.config.logging import LoggerLevels
+from proactor.config.logging import LoggingSettings
+from proactor.config.logging import RotatingFileHandlerSettings
+from proactor.config.mqtt import MQTTClient
+from proactor.config.paths import DEFAULT_BASE_DIR
+from proactor.config.paths import DEFAULT_BASE_NAME
+from proactor.config.paths import DEFAULT_LAYOUT_FILE
+from proactor.config.paths import DEFAULT_NAME
+from proactor.config.paths import DEFAULT_NAME_DIR
+from proactor.config.paths import Paths
+from proactor.config.proactor_settings import ProactorSettings
 
 DEFAULT_ENV_FILE = ".env"
 
@@ -41,6 +42,9 @@ __all__ = [
     "DEFAULT_NAME_DIR",
     "DEFAULT_LAYOUT_FILE",
     "Paths",
+
+    # proactor
+    "ProactorSettings",
 ]
